vector-store/requirements.txt
opensearch-py
python-dotenv
openai
langchain
langchain-community
langchain-openai
pypdf
psutil
boto3